fix FP32

+++ utils/metrics.py
@@ -86,7 +86,8 @@ class Evaluator():
             with torch.no_grad():
                 img_feat = model.encode_image(img).cpu()
                 image_feats = model.encode_image1(img)
-                token_features = model.img2text(img_feat.half().to(device))
+                # token_features = model.img2text(img_feat.half().to(device))
+                token_features = model.img2text(img_feat.to(device))
                 with autocast():
                     prompts = model.prompt_learner(token_features)
                     text_feature = model.text_encoder(prompts, model.prompt_learner.tokenized_prompts)
